Add Game entity
It will handle the game logic.
It provides a method to return the introduction message (delegated to
the dungeon) and another one to process the game commands.

The game data (for rooms and events) is loaded from yml files.

A main entry point is also provided to start the input loop, to print
the messages and to exit when it's done.